Register full mirrored jars with their network once.
Currently, full mirrored jars will never register themselves with their
network, meaning that if the server is restarted while a full mirrored
jar is in the world (or if a player places a full mirrored jar for the
first time since the server has started), the jar will be unable to
share essentia with the rest of its network.

This adds a check to register the jar even if it is full. The performance
impact should be minimal since we only check each jar once per run.